更新OrderFunction中的getCartTemp没有加入购物车功能

+++ src/com/BookStore/entity/Cart.java
@@ -6,13 +6,13 @@ public class Cart {
 	private String bookPhoto;
 	private int number;
 	private double price;
-	private int book_id;
+	private Long book_id;
 
-	public int getBook_id() {
+	public Long getBook_id() {
 		return book_id;
 	}
 
-	public void setBook_id(int book_id) {
+	public void setBook_id(Long book_id) {
 		this.book_id = book_id;
 	}
